backend/model/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==22.0.0
numpy
pandas
scikit-learn
networkx
requests
flask-cors
google-cloud-storage